feat: 캘린더 조회 API 구현

+++ module-domain/src/main/java/com/depromeet/member/Member.java
@@ -6,14 +6,14 @@ import lombok.Getter;
 @Getter
 public class Member {
     private Long id;
-    private Long goal;
+    private Integer goal;
     private String name;
     private String email;
     private MemberRole role;
     private String refreshToken;
 
     @Builder
-    public Member(Long id, Long goal, String name, String email, MemberRole role) {
+    public Member(Long id, Integer goal, String name, String email, MemberRole role) {
         this.id = id;
         this.goal = goal;
         this.name = name;
